Now it works like it should
Now it works!
I'd remove the ColorManger because i didn't get it work like I though. Instead I added a CodeItem which contains more information than the codeSnippet. I've added a initial array with colors to SettingsViewModel and method to calculate the color needed. For to get this work I've added some Extensions to Color and NSColor which allows you to create both from a HexString